pyproject.toml
[project]
name = "cli-pydantic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Tom Pollak", email = "tomp@graphcore.ai" }
]
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"
